requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
Pillow==10.2.0
numpy==1.26.3
opencv-python-headless==4.9.0.80
face_recognition==1.3.0
dlib==19.24.2
aiofiles==23.2.1
httpx==0.26.0
